getEventName bug fix
In some socket.io events, the code will be longer than "42", for
example: 427["someevent",{}]

The "7" refers to the id of the event. When that occurs, the
substring() method is messed up. This fix searches for the first double
quote, then searches for the second double quote AFTER the first one.