added getId() method
This method is used when socket.io does callback events. The beginning
of a fire and forget event starts with 42[ although with a callback
event it will include a number after the 42 and before the [ For
example: 425[ would have an id of 5. So a callback event is: 42{id}[

Then when doing a callback to the socket.io server the packet will
begin with 43{id}[] If the id=5, then the callback packet would be: 435[

I find this design is incredibly confusing! But that's just how
socket.io seems to work.