added callback method
I'm not sure if this is the right name for the method, but this is
incredibly useful. Socket.io has the ability to fire an event on a
client, then have that client respond back within the same request
scope. It does this with the use of IDs. Instead of the packet starting
with 42[ it will start with 42{id}[ For example, it may be: 425[ or
429[ This callback method will emit a packet back to the server with
43{id}[ If the server emitted 425[ then the callback would be 435[

This is extremely useful with node.js code like this:

io.sockets.connected[socketId].emit('getAccessPoints',{},function(access
Points){
	console.log(accessPoints);
});

Two things are happening here, the socket.io server is emitting the
"getAccessPoints" to the arduino board, then the arduino board is
responding BACk to the server with the variable "accessPoints"

So the server is sending something like:
425[{"getAccessPoints"}]

The arduino is responding with:
435[{"HP-Print-b1-LaserJet 200
color":"-51","Aslan":"-90","gibbon.net":"-68"}]