getEventPayload bug fix
It was throwing an error if an ID was passed with the event packet.
Instead of searching for the second double quote, I switched it to
search for the first comma which solves the problem.

This error doesn't come up if you have basic emit packet with something
like this:

42["eventname",{"hello":"world"}]

It would throw an error if the packet included an ID of say... 6, like
this:

426["eventname",{"hello":"world"}]

When the happens the count was off and it would use the first double
quote instead of the second one. By switching to the first comma, it
solves this problem.